requirements.txt
playwright>=1.45.0
python-dotenv>=1.0.0

